break: snippet syntax change
**BREAKING CHANGE**: The snippet syntax has been changed to use a new format.
From the old format:
tabstop: `$\d|$\{\d:[^\}]*\}`
visual: `${VISUAL}`
regex_group: `[[\d]]`

to the new format:
tabstop: `(?<!@)@\d|@\{\d\}|@\{\d:[^\}]*\}`
visual: `(?<!@)@{VISUAL}`
regex_group: `(?<!@)@\[\d\]`
literal `@`: `@@`

The tabstop syntax from `$` to `@` was needed to avoid conflicts with latex syntax as much as possible.
The regex group syntax was changed to standardize with the new tabstop syntax and for simplicity.
With now only one special character `@` for all tabstops, we only have to escape one character in the snippet text.
It should still be compatible with the old syntax if the version number is set to 1. Current version is 2.