feat: color and highlight brackets in math
when inside brackets it highlights the matching brackets even when not touching those.
matching brackets are colored with 3 different colors inside math to make it clearer which bracket matches with which like in code editors.

also made concealment and highlighting brackets live configurable and
used ts transpiler in the script that sends the settings to convert from
ts to js. This was done in there to since the typescript library is
large and it would slow down the parsing of the main module causing it
not to load.